pyproject.toml
[project]
name = "o3research"
dynamic = ["version"]
description = "O3 Deep Research tools"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "PyYAML>=6.0",
    "google-auth>=2.22.0",
    "google-auth-oauthlib>=1.1.0",
    "google-api-python-client>=2.99.0",
    "google-adk>=1.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["o3research*"]
exclude = ["node_modules*", "config*"]

[tool.setuptools]
include-package-data = true

[project.urls]
Homepage = "https://github.com/ADV-IT-Performance-Corp/ADV-ADK-Research-Hub"

[tool.setuptools.dynamic]
version = {file = "VERSION"}

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 88
